Update README.md
#Project Overview
This project involves the creation of a machine learning model designed to predict the likelihood of a heart attack in individuals based on their health data. The goal is to build a predictive tool that can assist in identifying at-risk patients by analyzing various health indicators. The model is built using Python and incorporates data preprocessing, feature engineering, and machine learning algorithms to achieve high prediction accuracy.